dark mode for documentation
