+++ website/src/repl/Editor.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useRef } from 'react';
 import { basicSetup } from 'codemirror';
 import { EditorView, keymap } from '@codemirror/view';
 import { defaultKeymap, indentWithTab } from '@codemirror/commands';
@@ -6,28 +6,12 @@ import { EditorState, Extension } from '@codemirror/state';
 import { javascript } from '@codemirror/lang-javascript';
 // TODO activate this when we have a dark mode
 import { oneDark } from '@codemirror/theme-one-dark';
+import useDarkMode from '../useDarkMode';
 
 type Props = {
   value: string;
   onChange: (value: string) => void;
 };
-
-function useDarkMode() {
-  const darkModeMediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
-  const [darkMode, setDarkMode] = useState(darkModeMediaQuery.matches);
-
-  useEffect(() => {
-    const handleChange = (e: MediaQueryListEvent) => {
-      setDarkMode(e.matches);
-    };
-    darkModeMediaQuery.addEventListener('change', handleChange);
-    return () => {
-      darkModeMediaQuery.removeEventListener('change', handleChange);
-    };
-  }, [darkModeMediaQuery]);
-
-  return darkMode;
-}
 
 export function Editor({ value, onChange }: Props): JSX.Element {
   const editor = useRef<HTMLDivElement>(null);
